[CHORE] UI Label, Button Extension 코드 변경(#8)
