--- Havit/Havit/Global/Extension/UILabel+.swift
@@ -1,36 +1,0 @@
-//
-//  UILabel+.swift
-//  Havit
-//
-//  Created by SHIN YOON AH on 2022/01/06.
-//
-
-import UIKit
-
-extension UILabel {
-    func addLabelSpacing(kernValue: Double = -0.6, paragraphValue: CGFloat = 4.0) {
-        if let labelText = text, labelText.count > 0 {
-            let attributedString = NSMutableAttributedString(string: labelText)
-            let paragraphStyle = NSMutableParagraphStyle()
-            paragraphStyle.lineSpacing = paragraphValue
-            attributedString.addAttribute(NSAttributedString.Key.kern, value: kernValue, range: NSRange(location: 0, length: attributedString.length - 1))
-            attributedString.addAttribute(NSAttributedString.Key.paragraphStyle, value: paragraphStyle, range: NSMakeRange(0, attributedString.length))
-            attributedText = attributedString
-            
-            if #available(iOS 14.0, *) {
-                lineBreakStrategy = .hangulWordPriority
-            } else {
-                lineBreakMode = .byWordWrapping
-            }
-        }
-    }
-    
-    func putSpecificColor(_ color: UIColor, range: String) {
-        if let labelText = text, labelText.count > 0 {
-            let attributedStr = NSMutableAttributedString(string: labelText)
-            attributedStr.addAttribute(.foregroundColor, value: color, range: (labelText as NSString).range(of: range))
-            attributedText = attributedStr
-        }
-    }
-}
-
