+++ Havit/Havit/Global/Extension/UILabel+Extension.swift
@@ -1,0 +1,30 @@
+//
+//  UILabel+Extension.swift
+//  Havit
+//
+//  Created by SHIN YOON AH on 2022/01/06.
+//
+
+import UIKit
+
+extension UILabel {
+    func addLabelSpacing(kernValue: Double = -0.6, lineSpacing: CGFloat = 4.0) {
+        if let labelText = text, labelText.count > 0 {
+            let paragraphStyle = NSMutableParagraphStyle()
+            paragraphStyle.lineSpacing = lineSpacing
+            attributedText = NSAttributedString(string: labelText,
+                                                attributes: [.kern: kernValue,
+                                                             .paragraphStyle: paragraphStyle])
+            lineBreakStrategy = .hangulWordPriority
+        }
+    }
+    
+    func applySpecificColor(_ color: UIColor, target: String) {
+        if let labelText = text, labelText.count > 0 {
+            let attributedStr = NSMutableAttributedString(string: labelText)
+            attributedStr.addAttribute(.foregroundColor, value: color, range: (labelText as NSString).range(of: target))
+            attributedText = attributedStr
+        }
+    }
+}
+
